Issues with extra / in api URL and was unable to pull until I changed the script
/usr/bin/defaults read "/Library/Preferences/com.jamfsoftware.jamf.plist" jss_url

Adds a / at the end

This was causing an issue with 
https://mydomain.jamfcloud.com//JSSResource/computers/udid/
creating an extra / before the JSSresource

Also I used md5 for my encryption (not sure why but I may have read that elsewhere)

This is my first attempt at a fork/commit and I don't know what I'm doing. So feel free to ignore this.